adds e2e test for hetzner
